+++ NEWS.md
@@ -1,6 +1,7 @@
 # cicerone 1.0.3
 
 - Add `tab` and `tab_id` arguments to trigger open tabs.
+- Add `is_id` argument to `step` and `highlight` method to allow using other selectors than `#id`, see [#7](https://github.com/JohnCoene/cicerone/issues/7)
 
 # cicerone 1.0.2